correct docker bash scripts for strict mode (#28)

+++ scripts/docker/network.bash
@@ -6,5 +6,7 @@ declare -r DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 . "${DIR%/*}/set_globals.bash"
 
 declare -r name="$PROJECT-net"
-docker network rm "$name"
+if [ $(docker network ls -qf name=$name) != "" ]; then
+    docker network rm "$name"
+fi
 docker network create --driver=bridge --subnet="$1" "$name"
